pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-linux"
version = "0.1.0"
description = "AI-powered package manager for Debian/Ubuntu that understands natural language"
readme = "README.md"
license = "BUSL-1.1"
authors = [
    {name = "Cortex Linux", email = "mike@cortexlinux.com"}
]
maintainers = [
    {name = "Mike Morgan", email = "mike@cortexlinux.com"}
]
keywords = [
    "linux",
    "package-manager",
    "ai",
    "llm",
    "natural-language",
    "apt",
    "debian",
    "ubuntu",
    "cli"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Installation/Setup",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    # LLM Provider APIs
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    # HTTP requests (for GitHub API, network operations)
    "requests>=2.32.4",
    # Configuration file parsing
    "pyyaml>=6.0.0",
    # Environment variable loading from .env files
    "python-dotenv>=1.0.0",
    # Encryption for environment variable secrets
    "cryptography>=42.0.0",
    # Terminal UI and formatting
    "rich>=13.0.0",
    # Type hints for older Python versions
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.3.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "isort>=5.0.0",
    "build>=0.10.0",
]
security = [
    "bandit>=1.7.0",
    "safety>=2.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]
all = [
    "cortex-linux[dev,security,docs]",
]

[project.scripts]
cortex = "cortex.cli:main"

[project.urls]
Homepage = "https://github.com/cortexlinux/cortex"
Documentation = "https://cortexlinux.com/docs"
Repository = "https://github.com/cortexlinux/cortex.git"
Issues = "https://github.com/cortexlinux/cortex/issues"
Discussions = "https://github.com/cortexlinux/cortex/discussions"
Changelog = "https://github.com/cortexlinux/cortex/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["cortex", "cortex.sandbox", "cortex.utils", "cortex.llm", "cortex.kernel_features", "cortex.kernel_features.ebpf", "cortex.tutor", "cortex.i18n"]
include-package-data = true

[tool.setuptools.package-data]
cortex = ["py.typed"]
"cortex.i18n" = ["locales/*.yaml"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.nox
    | \.tox
    | \.venv
    | venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "E402",   # module level import not at top (needed for sys.path manipulation)
    "E701",   # multiple statements on one line (test files)
    "E722",   # bare except (legacy code - fix incrementally)
    "E741",   # ambiguous variable name (l, O, I)
    "B007",   # unused loop control variable
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise from err (too many changes needed)
    "B905",   # zip without strict
    "F401",   # unused imports (some kept for API availability)
    "F841",   # local variable assigned but never used (argparse subparsers)
]

[tool.ruff.lint.isort]
known-first-party = ["cortex"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "build",
    "dist",
    "tests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::pytest.PytestConfigWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "requires_api: marks tests that require API keys",
    "asyncio: marks tests as async tests",
]

[tool.coverage.run]
source = ["cortex"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
fail_under = 55

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["cortex"]
skip = [".venv", "venv", "build", "dist"]
